+++ services/slb/upload_ca_certificate.go
@@ -95,9 +95,9 @@ type UploadCACertificateResponse struct {
 	Fingerprint       string `json:"Fingerprint" xml:"Fingerprint"`
 	ResourceGroupId   string `json:"ResourceGroupId" xml:"ResourceGroupId"`
 	CreateTime        string `json:"CreateTime" xml:"CreateTime"`
-	CreateTimeStamp   int    `json:"CreateTimeStamp" xml:"CreateTimeStamp"`
+	CreateTimeStamp   int64  `json:"CreateTimeStamp" xml:"CreateTimeStamp"`
 	ExpireTime        string `json:"ExpireTime" xml:"ExpireTime"`
-	ExpireTimeStamp   int    `json:"ExpireTimeStamp" xml:"ExpireTimeStamp"`
+	ExpireTimeStamp   int64  `json:"ExpireTimeStamp" xml:"ExpireTimeStamp"`
 	CommonName        string `json:"CommonName" xml:"CommonName"`
 }
